+++ loader/method.h
@@ -2,16 +2,17 @@
 #include<malloc.h>
 #include<util.h>
 #include<attributes.h>
+#include<stdint.h>
 
 typedef struct{
-    unsigned int access_flags;
-	unsigned int name_index;
-    unsigned int descriptor_index;
-	unsigned int attributes_count;
+    uint16_t access_flags;
+	uint16_t name_index;
+    uint16_t descriptor_index;
+	uint16_t attributes_count;
 	attribute_info *attributes;
 }method_info;
 
-extern unsigned int methodCount;
+extern uint16_t methodCount;
 extern method_info *methods;
 
 void parseMethods(FILE *r);
